+++ QLMarkdown/Settings+render.swift
@@ -223,7 +223,16 @@ extension Settings {
                 os_log("Could not enable markdown `heads` extension!", log: OSLog.rendering, type: .error)
             }
         }
-        
+
+        if self.definitionListExtension {
+            if let ext = cmark_find_syntax_extension("definitionlist") {
+                cmark_parser_attach_syntax_extension(parser, ext)
+                os_log("Enabled markdown `definitionlist` extension.", log: OSLog.rendering, type: .debug)
+            } else {
+                os_log("Could not enable markdown `definitionlist` extension!", log: OSLog.rendering, type: .error)
+            }
+        }
+
         if self.highlightExtension {
             if let ext = cmark_find_syntax_extension("highlight") {
                 cmark_parser_attach_syntax_extension(parser, ext)
@@ -247,6 +256,19 @@ extension Settings {
                     type: .debug)
             } else {
                 os_log("Could not enable markdown `sub` extension!", log: OSLog.rendering, type: .error)
+            }
+        }
+
+        if self.alertExtension {
+            if let ext = cmark_find_syntax_extension("alert") {
+                cmark_parser_attach_syntax_extension(parser, ext)
+
+                os_log(
+                    "Enabled markdown `alert` extension.",
+                    log: OSLog.rendering,
+                    type: .debug)
+            } else {
+                os_log("Could not enable markdown `alert` extension!", log: OSLog.rendering, type: .error)
             }
         }
         
@@ -709,6 +731,13 @@ table.debug td {
             html_debug += "off"
         }
         html_debug += "</td></tr>\n"
+        html_debug += "<tr><td>alert extension</td><td>"
+        if self.alertExtension {
+            html_debug += "on " + (cmark_find_syntax_extension("alert") == nil ? " (NOT AVAILABLE" : "")
+        } else {
+            html_debug += "off"
+        }
+        html_debug += "</td></tr>\n"
         
         html_debug += "<tr><td>table extension</td><td>"
         if self.tableExtension {
@@ -844,11 +873,9 @@ table.debug td {
             
         // Custom Style applies in both modes; only the bundled GitHub default.css is
         // specific to Markdown (non-source) mode.
-        let css = (self.customCSSFetched ? self.customCSSCode : self.getCustomCSSCode()) ?? ""
-        css_doc_extended = formatCSS(css)
-        if !self.renderAsCode, css_doc_extended.isEmpty || !self.customCSSOverride {
-            css_doc += formatCSS(getBundleContents(forResource: "default", ofType: "css"))
-        }
+        let css = self.getAppliedCSS()
+        css_doc_extended = formatCSS(css.custom)
+        css_doc += formatCSS(css.bundled)
             
         var css_highlight: String = ""
         if self.renderAsCode {